refactor: clean up signup component
