+++ types/auth.ts
@@ -10,3 +10,5 @@ export const loginSchema = z.object({
   email: z.string().email(),
   password: z.string().min(6),
 });
+
+export type signupSchema = z.infer<typeof signupSchema>;
